MPEG 1/2: Eliminate the usage of M2VParser.GetRealSequenceHeader()
The mpeg 1/2 video packetizer already does the job of setting private data.
Besides, the function does not seem to return a sequence header reliably,
which is why the private data is malformed even when the packetizer calls
it before it calls remove_stuffing_bytes_and_handle_sequence_headers().
(Maybe that is why there is the remark "BE VERY CAREFUL WITH THIS CALL".)