Silences the `has no symbols` warnings on darwin
Building a static library using `ar cr` and `ranlib` on darwin produces
`has no symbols` warnings when trying to add a symbol table to an empty
archive file.

This can be silenced by passing the `-S` option to `ar` (prevents the
creation of a symbol table when creating the archive file) and the
`-no_warning_for_no_symbols` option to `ranlib` when adding the symbol
table.

Unfortunately the `-no_warning_for_no_symbols` is only supported on
darwin and not on linux and thus we need to pass it conditionally.