More explicit control over use of QtDbus
Relates to #2886

I'm packaging mkvtoolnix for an environment that doesn't use dbus
and I found myself wanting more explicit control of whether or not
dbus is enabled.

This patch adds --enable-dbus (default) and --disable-dbus configure
options. If we ask for it to be enabled, the absence of it becomes
a fatal configuration error.

If dbus is installed on the system, and any reason a person wants to
exclude that functionality from mkvtoolnix. That is now possible too.